Move hashable concept to vast/hash
To simplify the directory structure and avoid confusion with native
concepts, we move the "hashable concept" to its own top-level directory
within the libvast source tree.
