+++ libvast/src/system/passive_partition.cpp
@@ -13,7 +13,6 @@
 #include "vast/address_synopsis.hpp"
 #include "vast/aliases.hpp"
 #include "vast/chunk.hpp"
-#include "vast/concept/hashable/xxhash.hpp"
 #include "vast/concept/printable/to_string.hpp"
 #include "vast/concept/printable/vast/expression.hpp"
 #include "vast/concept/printable/vast/table_slice.hpp"
@@ -27,6 +26,7 @@
 #include "vast/fbs/partition.hpp"
 #include "vast/fbs/utils.hpp"
 #include "vast/fbs/uuid.hpp"
+#include "vast/hash/xxhash.hpp"
 #include "vast/ids.hpp"
 #include "vast/legacy_type.hpp"
 #include "vast/logger.hpp"
