+++ libvast/test/bloom_filter_synopsis.cpp
@@ -10,8 +10,8 @@
 
 #include "vast/bloom_filter_synopsis.hpp"
 
-#include "vast/concept/hashable/hash_append.hpp"
-#include "vast/concept/hashable/xxhash.hpp"
+#include "vast/hash/hash_append.hpp"
+#include "vast/hash/xxhash.hpp"
 #include "vast/si_literals.hpp"
 #include "vast/test/synopsis.hpp"
 #include "vast/test/test.hpp"
