+++ libvast/vast/data.hpp
@@ -10,14 +10,14 @@
 
 #include "vast/address.hpp"
 #include "vast/aliases.hpp"
-#include "vast/concept/hashable/uhash.hpp"
-#include "vast/concept/hashable/xxhash.hpp"
 #include "vast/concept/printable/print.hpp"
 #include "vast/concepts.hpp"
 #include "vast/data/integer.hpp"
 #include "vast/defaults.hpp"
 #include "vast/detail/operators.hpp"
 #include "vast/detail/type_traits.hpp"
+#include "vast/hash/uhash.hpp"
+#include "vast/hash/xxhash.hpp"
 #include "vast/legacy_type.hpp"
 #include "vast/offset.hpp"
 #include "vast/pattern.hpp"
