+++ libvast/vast/hash/crc.hpp
