+++ libvast/vast/hash/hasher.hpp
