+++ libvast/vast/hash/sha1.hpp
