+++ libvast/vast/hash/xxhash.hpp
