+++ libvast/vast/legacy_type.hpp
@@ -12,14 +12,14 @@
 
 #include "vast/aliases.hpp"
 #include "vast/attribute.hpp"
-#include "vast/concept/hashable/hash.hpp"
-#include "vast/concept/hashable/xxhash.hpp"
 #include "vast/concept/printable/print.hpp"
 #include "vast/detail/assert.hpp"
 #include "vast/detail/operators.hpp"
 #include "vast/detail/range.hpp"
 #include "vast/detail/stack_vector.hpp"
 #include "vast/detail/type_traits.hpp"
+#include "vast/hash/hash.hpp"
+#include "vast/hash/xxhash.hpp"
 #include "vast/offset.hpp"
 #include "vast/operator.hpp"
 #include "vast/time.hpp"
